Check for errors and warnings in build script as well (#440)
